피드백 및 databinding 일부 적용

+++ app/src/main/java/com/example/myapplication/model/TrackingLog.kt
@@ -1,7 +1,6 @@
 package com.example.myapplication.model
 
-import androidx.room.ColumnInfo
-import androidx.room.Embedded
+import androidx.databinding.BaseObservable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import java.util.*
@@ -14,5 +13,5 @@ data class TrackingLog(
     @PrimaryKey(autoGenerate = true) val trackingId: Int=0,
     val trackingStartTime: Date, //기록 시작한 시각
     val trackingEndTime: Date, //기록 종료한 시각
-    val trackingDistance: Double
+    val trackingDistance: Int
 )
